src/main/java/com/intel/databackend/api/inquiry/advanced/aggregations/AggregationResult.java
/**
 * Copyright (c) 2015 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.intel.databackend.api.inquiry.advanced.aggregations;

import com.intel.databackend.datastructures.AdvancedComponent;

public class AggregationResult {

    private Double min;
    private Double max;
    private Double sum;
    private Double sumOfSquares;
    private long count;

    public AggregationResult() {
    }

    public Double getMin() {
        return min;
    }

    public void setMin(Double min) {
        this.min = min;
    }

    public Double getMax() {
        return max;
    }

    public void setMax(Double max) {
        this.max = max;
    }

    public Double getSum() {
        return sum;
    }

    public void setSum(Double sum) {
        this.sum = sum;
    }

    public Double getSumOfSquares() {
        return sumOfSquares;
    }

    public void setSumOfSquares(Double sumOfSquares) {
        this.sumOfSquares = sumOfSquares;
    }

    public long getCount() {
        return count;
    }

    public void setCount(long count) {
        this.count = count;
    }

    public void addToComponent(AdvancedComponent component) {
        component.setMax(getMax());
        component.setMin(getMin());
        component.setCount(getCount());
        component.setSum(getSum());
        component.setSumOfSquares(getSumOfSquares());
    }
}
